requirements.txt
eve
eve-swagger
click
PyYAML
